[Refactor] global exception으로 변경
