+++ project/src/main/java/org/example/fourtreesproject/exception/custom/InvalidBidException.java
@@ -1,4 +1,4 @@
-package org.example.fourtreesproject.bid.exception.customs;
+package org.example.fourtreesproject.exception.custom;
 
 
 import lombok.Getter;
