+++ project/src/main/java/org/example/fourtreesproject/orders/service/OrdersService.java
@@ -16,15 +16,15 @@ import org.example.fourtreesproject.coupon.model.UserCoupon;
 import org.example.fourtreesproject.coupon.repository.UserCouponRepository;
 import org.example.fourtreesproject.delivery.model.DeliveryAddress;
 import org.example.fourtreesproject.delivery.model.response.DeliveryAddressResponse;
+import org.example.fourtreesproject.exception.custom.InvalidOrderException;
 import org.example.fourtreesproject.groupbuy.model.entity.GroupBuy;
 import org.example.fourtreesproject.groupbuy.repository.GroupBuyRepository;
-import org.example.fourtreesproject.orders.exception.custom.InvalidOrderException;
 import org.example.fourtreesproject.orders.model.entity.Orders;
 import org.example.fourtreesproject.orders.model.response.OrderPageResponse;
 import org.example.fourtreesproject.orders.model.response.OrdersListResponse;
 import org.example.fourtreesproject.orders.repository.OrdersRepository;
 import org.example.fourtreesproject.product.model.entity.Product;
-import org.example.fourtreesproject.user.exception.custom.InvalidUserException;
+import org.example.fourtreesproject.exception.custom.InvalidUserException;
 import org.example.fourtreesproject.user.model.entity.User;
 import org.example.fourtreesproject.user.model.entity.UserDetail;
 import org.example.fourtreesproject.user.model.response.UserCouponResponse;
